% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRates.R
\docType{methods}
\name{compareRates,pulseTDmodel-method}
\alias{compareRates,pulseTDmodel-method}
\title{compareRates-Method}
\usage{
\S4method{compareRates}{pulseTDmodel}(object1, object2)
}
\arguments{
\item{object1}{An object of class 'pulseTDmodel'}

\item{object1}{An object of class 'pulseTDmodel'}
}
\value{
A list
}
\description{
compare result.
}
