% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{compareRates}
\alias{compareRates}
\title{compareRates Method}
\usage{
compareRates(object1, object2)
}
\arguments{
\item{object1}{An object of class 'pulseTDmodel'}

\item{object1}{An object of class 'pulseTDmodel'}
}
\value{
A list
}
\description{
compare result.
}
\examples{
data('pulseRates', package='pulseTD')
}
