% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionParams.R
\name{correctionParams}
\alias{correctionParams}
\title{Correct the parameters of the pulse model}
\usage{
correctionParams(object)
}
\arguments{
\item{object}{a 'pulseTDmodel' that has been calculated with \code{\link{estimateParams}}}
}
\value{
a 'pulseTDmodel' that has been modified
}
\description{
It is used to correct pulse model parameters.
During the estimation of parameters in \code{\link{estimateParams}},
some NA will occur, which may be caused by random initial values.
Therefore, some NA values can be estimated from the new ones.
}
\examples{
data('pulseRates', package='pulseTD')
pulseRates_correct = correctionParams(pulseRates)
pulseRates_correct@fitfailure
}
