% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\docType{methods}
\name{getParams,pulseTDmodel-method}
\alias{getParams,pulseTDmodel-method}
\title{getParams-method}
\usage{
\S4method{getParams}{pulseTDmodel}(object, stage, genename = NULL)
}
\arguments{
\item{object}{An object of class 'pulseTDmodel'}

\item{stage}{A character such as transcription, processing or degradation for rates.}

\item{genename}{A vector of gene names, default is NULL.}
}
\value{
A numeric matrix containing the rates
}
\description{
It is used to obtain transcription, processing and degradation rates.
}
