% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRates.R
\docType{methods}
\name{getRates,pulseTDmodel-method}
\alias{getRates,pulseTDmodel-method}
\title{getRates-method}
\usage{
\S4method{getRates}{pulseTDmodel}(object, stage, timevector = NULL,
  genename = NULL)
}
\arguments{
\item{object}{An object of class 'pulseTDmodel'}

\item{stage}{A character such as transcription, processing or degradation for rates.}

\item{timevector}{A vector of times, default is NULL.}

\item{genename}{A vector of gene names, default is NULL.}
}
\value{
A numeric matrix containing the rates
}
\description{
It is used to obtain transcription, processing and degradation rates.
}
