% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapRates.R
\docType{methods}
\name{heatmapParams,pulseTDmodel-method}
\alias{heatmapParams,pulseTDmodel-method}
\title{heatmapParams-Method}
\usage{
\S4method{heatmapParams}{pulseTDmodel}(object, genename = NULL)
}
\arguments{
\item{object}{An object of class 'pulseTDmodel'}

\item{genename}{The name of the gene you want to draw.}
}
\value{
No return value
}
\description{
Draw the image of the fitting result.
}
