% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic.R
\name{heatmapParams}
\alias{heatmapParams}
\title{heatmapParams Method}
\usage{
heatmapParams(object, genename = NULL)
}
\arguments{
\item{object}{a pulseTDmodel that has been calculated with \code{\link{estimateParams}}}

\item{genename}{The name of the gene to be drawn}
}
\value{
There is no return value.
}
\description{
It is used to draw a paramaters image.
}
\examples{
data('pulseRates', package='pulseTD')
heatmapParams(pulseRates, 1:10)
}
