% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseModel.R
\name{integrateAC}
\alias{integrateAC}
\title{integral function}
\usage{
integrateAC(xita, timegrid)
}
\arguments{
\item{xita}{Six parameters of the pulse function :h0, h1, h2, t1, t2, beta}

\item{timegrid}{a vector or point of time}
}
\value{
a vector or point of exon expression value
}
\description{
Calculate the expression values at different time points.
}
