% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalFactors.R
\name{normFactors}
\alias{normFactors}
\title{esitmate the global normalization factor}
\usage{
normFactors(labexon, totexon, TimeGrid, tL, batch = 2, samumber = 50,
  iternumber = 50)
}
\arguments{
\item{labexon}{Exon expression value of 4sU label RNA}

\item{totexon}{Exon expression value of total RNA}

\item{TimeGrid}{a vector  of times}

\item{tL}{label time}

\item{batch}{Number of genes evaluated simultaneously}

\item{samumber}{Number of samples randomly drawn per iteration}

\item{iternumber}{Number of iterations per sample}
}
\value{
a vector of the global normalization factor
}
\description{
It is used to the global normalization factor.
}
