% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRates.R
\docType{methods}
\name{plotRates,pulseTDmodel-method}
\alias{plotRates,pulseTDmodel-method}
\title{plotRate-Method}
\usage{
\S4method{plotRates}{pulseTDmodel}(object, genename, predict = FALSE)
}
\arguments{
\item{object}{An object of class 'pulseTDmodel'}

\item{genename}{The name of the gene you want to draw.}

\item{predict}{The time interval you want to draw.}
}
\value{
No return value
}
\description{
Draw the image of the fitting result.
}
