% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdatadoc.R
\name{preExp}
\alias{preExp}
\title{A gene list that stores predicted values of gene expression and a 0.95 confidence interval.}
\format{A list}
\description{
A gene list that stores predicted values of gene expression and a 0.95 confidence interval.
}
