% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdatadoc.R
\name{pulseRates}
\alias{pulseRates}
\title{A pulseTDModel Rdata file containing 100 gene expression data,
transcriptional dynamic data already related parameters.}
\format{A pulseTD model}
\description{
A pulseTDModel Rdata file containing 100 gene expression data,
transcriptional dynamic data already related parameters.
}
