% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdatadoc.R
\name{rpkmSim}
\alias{rpkmSim}
\title{A simulated gene expressing Rdata file that contains expression data for total RNA, precursor RNA, and 4sU-labeled RNA.}
\format{A list}
\description{
A simulated gene expressing Rdata file that contains expression data for total RNA, precursor RNA, and 4sU-labeled RNA.
}
