% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptionParams.R
\name{splitA6Params}
\alias{splitA6Params}
\title{esitmate transcription parameters}
\usage{
splitA6Params(TT, TL, tL, TimeGrid, w1, w2)
}
\arguments{
\item{TL}{Exon expression value of 4sU label RNA}

\item{tL}{label time}

\item{TimeGrid}{a vector  of times}
}
\value{
a vector of transcription parameters,contains pulse parameters and fitting error values
}
\description{
It is used to solve the transcription parameters, including the error function and the optimization function.
}
