% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradationParams.R
\name{splitB4Params}
\alias{splitB4Params}
\title{esitmate degradation parameters}
\usage{
splitB4Params(TT, pre_a, TimeGrid)
}
\arguments{
\item{TT}{exon expression value of total RNA}

\item{pre_a}{Parameter vector of transcription pulse function}

\item{TimeGrid}{a vector  of times}
}
\value{
a vector of degradation parameters,contains pulse parameters and fitting error values
}
\description{
It is used to solve the degradation parameters, including the error function and the optimization function.
}
