% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingParams.R
\name{splitC5Params}
\alias{splitC5Params}
\title{esitmate processing parameters}
\usage{
splitC5Params(PT, pre_a, TimeGrid)
}
\arguments{
\item{PT}{Intron expression value of total RNA}

\item{pre_a}{Parameter vector of transcription pulse function}

\item{TimeGrid}{a vector of times}
}
\value{
a vector of processing parameters,contains pulse parameters and fitting error values
}
\description{
It is used to solve the processing parameters, including the error function and the optimization function.
}
